Release to prepare for Q1 2022 release
CustomerInfo
Manual Registration fix

Fix intermitted failures in generic_numCreditsRows_NotAllowed
by removing .parallel processing of streams from non threadsafe classes
