+++ pom.xml
@@ -45,9 +45,9 @@
         <java.version>11</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <ob-aspsp.version>1.7.1</ob-aspsp.version>
-        <ob-common.version>1.3.3</ob-common.version>
-        <ob-clients.version>1.3.2</ob-clients.version>
+        <ob-aspsp.version>1.7.3</ob-aspsp.version>
+        <ob-common.version>1.3.4</ob-common.version>
+        <ob-clients.version>1.3.3</ob-clients.version>
         <!-- others -->
         <commons-csv.version>1.7</commons-csv.version>
         <springboot-test.version>2.1.5.RELEASE</springboot-test.version>
